requirements.txt
flask==3.0.0
google-cloud-texttospeech>=2.29.0
gunicorn==21.2.0

